Frontend_streamlit/requirements.txt
streamlit == 1.19.0
requests == 2.28.2
numpy == 1.24.2
Pillow == 9.4.0
s3fs == 2023.3.0
protobuf==3.20.0
numpy == 1.24.2
pandas == 1.5.3
plotly-express == 0.4.1
folium == 0.14.0
streamlit-folium == 0.11.1
